[svelte] Fix sync of pulse animations
